+++ docker-compose.yml
@@ -14,14 +14,23 @@ services:
       - minio_storage:/data
 
   postgres:
-    image: postgres
+    image: postgres:15.4-alpine
     env_file:
       - .env
     ports:
       - "5432:5432"
     volumes:
       - pgdata:/var/lib/postgresql/data
- 
+
+  redis:
+    image: redis:7.2-alpine
+    ports:
+      - "6379:6379"
+    command: redis-server --save 60 1 --loglevel warning
+    volumes: 
+      - redis:/data
+
 volumes:
   pgdata:
+  redis:
   minio_storage:
