+++ package.json
@@ -17,7 +17,9 @@
     "@prisma/client": "^5.2.0",
     "@types/formidable": "^3.4.2",
     "@types/uuid": "^9.0.3",
+    "bullmq": "^4.9.0",
     "formidable": "^3.5.1",
+    "ioredis": "^5.3.2",
     "masto": "^6.2.0",
     "minio": "^7.1.2",
     "nuxt": "^3.7.0",
