+++ server/api/threads/[id]/schedule.post.ts
@@ -1,0 +1,46 @@
+import { Thread } from '~/models/models'
+import { Worker } from 'bullmq'
+import { publish } from '~/services/publication-service'
+import { connection, queue } from '~/utils/queue'
+import { prisma } from '~/prisma/db'
+
+export default defineEventHandler(async (event: any) => {
+    const threadId = parseInt(event.context.params.id) as number
+    const data: any = await readBody(event)
+
+    console.log(`POST /api/threads/${threadId}/schedule`)
+
+    const now = new Date()
+    const scheduledAt = now.setSeconds(now.getSeconds() + 30)
+    const delay = scheduledAt - Date.now()
+
+    const threadData = await prisma.thread.findFirst({ where: { id: threadId } })
+    if (!threadData) {
+        throw new Error(`Could not publish thread with ID ${threadId} because it does not exist.`)
+    }
+    await prisma.thread.update({
+        where: { id: threadId },
+        data: { scheduledAt: data.scheduledAt }
+    })
+
+    await queue.add(`thread-${threadId}`, { threadId: threadId }, { delay });
+
+    const worker = new Worker('thread-schedules', async (job) => {
+        try {
+            const thread: Thread = await publish(threadId)
+            return thread
+        } catch (error) {
+            return {
+                status: 'error',
+                data: error
+            }
+        }
+    }, { connection })
+    worker.on('completed', job => {
+        console.log(`${job.id} has completed!`);
+    });
+
+    worker.on('failed', (job, err) => {
+        console.log(`${job?.id} has failed with ${err.message}`);
+    });
+})
